Update engine v070 (#23)
* update anchore-engine to v0.7.0

* update vulnerable deps

* Fix TypeError introduced in GitHub core module
The GitHub core module expects all console output to be passed as a
string. In the findFiles function, core.debug was being called with a
json object as the argument causing a TypeError. Using toString() on all
arguments passed to core.debug fixes this issue.

* ignore dev deps when using npm audit in CI
ESlint has a dep with an unfixed vulnerability. Since this is a dev
dependency, it does not need to be checked in CI for vulns.

* remove debug statement and make image-reference required again

* fix test mocks & async promise error
jest requires mocking the entire fs module if you want to overwrite a
specific function. However, this breaks the action itself. So instead of
mocking the entire module, you can use jest.spyOn to only mock specific
functions from the module, then mockRestore() to get the original
functionality back.

JSON.stringfy was not returning a promise which caused the aysnc jobs to
fail. By wrapping this in try/catch you can ensure the promise is
captured and fix async promise errors.

Signed-off-by: Brady Todhunter <bradyt@anchore.com>