fix: updated semver version
While updating the action with a new grype version npm run audit
fails because of an issue with semver being vulnerable.
https://github.com/anchore/scan-action/actions/runs/6479148648/job/17592137397?pr=236

The upgrade of semver is also failing, so the action is stuck with
an old grype version.

This commit has an upgraded semver version and a recreated index.js

Signed-off-by: Gian Carlo Pace <giancarlo.pace@gmail.com>